+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 1.1.3 - 2024-10-02
+- Added Live example to documentation page
+- Improved featureList api, now you can get the list of features as `String` or `JSON` format.
+
+### 1.1.1 - 2024-10-02
+- Added Options to initialize function to enable/disable localFlags globally
+
 ### 1.1.0 - 2024-10-02
 - Update API to group API into `localFlags` global variable